# Docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: BOSSNYUMBA Property Management API
  description: |
    Comprehensive REST API for the BOSSNYUMBA multi-tenant property management SaaS platform.
    
    ## Authentication
    All endpoints (except `/auth/login`, `/auth/register`, `/auth/forgot-password`, `/auth/demo-users`,
    and `/payments/mpesa/callback`) require a valid JWT Bearer token in the `Authorization` header.
    
    ## Multi-Tenancy
    All data is scoped to the authenticated user's tenant. The tenant context is derived from the JWT token.
    
    ## Pagination
    List endpoints support pagination via `page` and `pageSize` query parameters.
    Responses include a `pagination` object with `page`, `pageSize`, `totalItems`, `totalPages`,
    `hasNextPage`, and `hasPreviousPage` fields.
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable description"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: BOSSNYUMBA Engineering
    email: engineering@bossnyumba.com
  license:
    name: Proprietary
    url: https://bossnyumba.com/terms
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.bossnyumba.com/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /properties/{id}/units:
    get:
      tags: [Properties]
      summary: List units for a property
      operationId: listPropertyUnits
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, OCCUPIED, RESERVED, MAINTENANCE]
      responses:
        '200':
          description: List of units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /units:
    get:
      tags: [Units]
      summary: List all units
      operationId: listUnits
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, OCCUPIED, RESERVED, MAINTENANCE]
        - name: type
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [unitNumber, floor, rentAmount, status, createdAt]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      tags: [Units]
      summary: Create a unit
      operationId: createUnit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitRequest'
      responses:
        '201':
          description: Unit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '409':
          description: Duplicate unit number in property
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /units/{id}:
    get:
      tags: [Units]
      summary: Get unit by ID (includes current lease/tenant)
      operationId: getUnit
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Unit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Units]
      summary: Update unit
      operationId: updateUnit
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUnitRequest'
      responses:
        '200':
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Units]
      summary: Soft-delete unit
      operationId: deleteUnit
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Unit deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /units/{id}/status:
    put:
      tags: [Units]
      summary: Update unit status
      operationId: updateUnitStatus
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [AVAILABLE, OCCUPIED, RESERVED, MAINTENANCE]
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ===========================================================================
  # CUSTOMER ENDPOINTS
  # ===========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from POST /auth/login.
        Include as: `Authorization: Bearer <token>`
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource unique identifier
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean

    # =========================================================================
    # Auth schemas
    # =========================================================================
...
    CreateUnitRequest:
      type: object
      required: [propertyId, unitNumber, rentAmount]
      properties:
        propertyId:
          type: string
        unitNumber:
          type: string
          example: A101
        floor:
          type: integer
        type:
          type: string
          enum: [STUDIO, ONE_BEDROOM, TWO_BEDROOM, THREE_BEDROOM, PENTHOUSE, OFFICE, SHOP]
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, RESERVED, MAINTENANCE]
          default: AVAILABLE
        bedrooms:
          type: integer
        bathrooms:
          type: number
        squareMeters:
          type: number
        rentAmount:
          type: number
        depositAmount:
          type: number
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
    UpdateUnitRequest:
      type: object
      properties:
        unitNumber:
          type: string
        floor:
          type: integer
        type:
          type: string
        status:
          type: string
        bedrooms:
          type: integer
        bathrooms:
          type: number
        squareMeters:
          type: number
        rentAmount:
          type: number
        depositAmount:
          type: number
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string

    # =========================================================================
    # Customer schemas
    # =========================================================================
...